refactor: unify AudioRecorder + RawAudioCapture into UnifiedAudioEngine (#39)
Replace two separate audio engines (AudioRecorder wrapping WhisperKit's
AudioProcessor, RawAudioCapture using native AVAudioEngine) with a single
UnifiedAudioEngine based on native AVAudioEngine.

Key changes:
- isEngineRunning is now a computed property (engine.isRunning) — fixes #38
- Sample gating: processBuffer() only accumulates audio when isRecording=true,
  eliminating idle sample accumulation bug
- DictationCoordinator simplified: 3 start paths → 2, 2 stop branches → 1,
  4 Combine subscriptions → 2, ~930 lines removed
- WhisperKit decoupled from audio capture — only used for transcription now

Closes #39, fixes #38
